Move examples to kotlin asset folder; add kotlin toolchain
