+++ velocity/example-plugin/src/main/kotlin/com/example/KotlinExamplePlugin.kt
@@ -1,16 +1,27 @@
 package com.example
 
-import com.hypixel.hytale.server.core.plugin.JavaPlugin
-import com.hypixel.hytale.server.core.plugin.JavaPluginInit
+import com.google.inject.Inject
+import com.velocitypowered.api.event.Subscribe
+import com.velocitypowered.api.event.proxy.ProxyInitializeEvent
+import com.velocitypowered.api.event.proxy.ProxyShutdownEvent
+import com.velocitypowered.api.plugin.Plugin
 import dev.faststats.ErrorTracker
 import dev.faststats.data.Metric
-import dev.faststats.hytale.HytaleContext
+import dev.faststats.velocity.VelocityContext
 import java.util.concurrent.atomic.AtomicInteger
 
-class ExamplePlugin(init: JavaPluginInit) : JavaPlugin(init) {
+@Plugin(
+    id = "example",
+    name = "Example Plugin",
+    version = "1.0.0",
+    url = "https://example.com",
+    authors = ["Your Name"],
+)
+class KotlinExamplePlugin @Inject constructor(contextBuilder: VelocityContext.Builder) {
     private val gameCount = AtomicInteger()
 
-    private val context = HytaleContext.Factory(this, "YOUR_TOKEN_HERE")
+    private val context = contextBuilder
+        .token("YOUR_TOKEN_HERE")
         .errorTrackerService(ERROR_TRACKER)
         // .metrics(Metrics.Factory::create) // Define a minimal metrics instance without any custom metrics
         .metrics { factory ->
@@ -27,11 +38,13 @@ class ExamplePlugin(init: JavaPluginInit) : JavaPlugin(init) {
         }
         .create()
 
-    override fun setup() {
+    @Subscribe
+    fun onProxyInitialize(event: ProxyInitializeEvent) {
         context.ready() // start metrics and errors submission
     }
 
-    override fun shutdown() {
+    @Subscribe
+    fun onProxyStop(event: ProxyShutdownEvent) {
         context.shutdown() // safely shut down configured services
     }
